refactor and file cleanups (#49)
* Merge into main (#5)

* Update README.md (#1)

* Create nodejs.yml (#2)

* Basic bot setup (#3)

* Setup (#4)

* Basic bot setup

* Add procfile and fake test


* refactor start.js

* refactor endConditions

* refactor countdown, start and nickname auto replies

* general refactor of all files
