+++ commands/challenge.js
@@ -1,4 +1,5 @@
-const { checkChallenge } = require('../utils/timeInterrupt');
+const { botReplies } = require('../data/shameReplies');
+const { checkChallenge } = require('../utils/challengeData');
 const { usersArray }  = require('./start');
 
 function ifChallenge(message){
@@ -11,7 +12,7 @@ function ifChallenge(message){
         checkChallenge(message, user);       
       }
       if(message.author.id === user.userId && user.mode !== 'lockdown'){
-        message.reply('you are not in a mode which requires a challenge before exiting. To end your session early enter command --exit');      
+        message.reply(botReplies.noChallenge());      
       }
     }
   }
