--- stretch/setupWizard.js
@@ -1,1 +1,0 @@
-
