+++ utils/challengeData.js
@@ -1,5 +1,3 @@
-// used with the stop command, interrupts a timeout and ends the altered user state
-
 const { botReplies } = require('../data/shameReplies');
 
 const wordCount = 50;
