+++ utils/checkOwner.js
@@ -1,0 +1,9 @@
+const isUserOwner = (message) => {
+  const { member, guild } = message;
+  return member === guild.owner ? true : false;
+};
+
+
+module.exports = {
+  isUserOwner,
+};
