--- utils/checkRoleStatus.js
@@ -1,6 +1,0 @@
-const isBotRoleHigher = (message) => {
-  const { member } = message;
-  return member.manageable;
-};
-
-module.exports = { isBotRoleHigher };
