+++ utils/endConditions.js
@@ -7,38 +7,26 @@ const cleanUp = async (usersArray) => {
   for(let i = 0; i < usersArray.length; i++){
     const user = usersArray[i];
 	
-    // TODO maybe refactor here to divide up if blocks
     if(user.endTime < Date.now() || !user.isActive){ 
 	
-      if(!user.isActive){
-        restoreNickname(user, user.member);
+      restoreNickname(user, user.member);
 	
-        if(user.mode !== 'shame') {
-          removeChannelOverwrites(user);
-          deleteChannel(user.newChannel);
-        } else {
-          if(user.userSetTimer){
-            user.botTimerMessage.delete();
-          }
+      if(user.mode !== 'shame') {
+        removeChannelOverwrites(user);
+        deleteChannel(user.newChannel);
+      } else {
+        if(user.userSetTimer){
+          user.botTimerMessage.delete();
         }
       }
 	
-      if(user.isActive && !user.member.guild.owner){
-        user.originalChannel.send(botReplies.timerEnded(user.userId));
-			
-        restoreNickname(user, user.member);
-	
-        if(user.mode !== 'shame') {
-          removeChannelOverwrites(user);
-          deleteChannel(user.newChannel);
-        } else {
-          if(user.userSetTimer){
-            user.botTimerMessage.delete();
-          }
-        }
+      if(user.isActive){
+        setTimeout(() => {
+
+          user.originalChannel.send(botReplies.timerEnded(user.userId));
+          
+        }, 1000);
       }
-	
-      if(user.isActive && user.member.guild.owner)user.originalChannel.send(botReplies.timerEnded(user.userId)); 
 	
       usersArray.splice(i, 1);
       i--;
