+++ utils/parseTime.js
@@ -1,7 +1,3 @@
-// takes user time input and parses it to milliseconds
-
-// ideally, enforcing a 00:00 format might be the most consistently parseable
-
 function parseTime(stringTime){
   const hrs = Number(stringTime.split(':')[0]);
   const mins = Number(stringTime.split(':')[1]);
